Move latest-import from user to api
Also, some documentation updates
