+++ docs/conf.py
@@ -26,8 +26,8 @@ import shlex
 # If your documentation needs a minimal Sphinx version, state it here.
 #needs_sphinx = '1.0'
 
-sys.path.insert(0, os.path.abspath('..'))
-import webserver
+sys.path.append(os.path.abspath('..'))
+from listenbrainz import webserver
 
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
